Refactor web service to use flask blueprint
