+++ services/web/project/tests/base.py
@@ -1,7 +1,8 @@
 from flask_testing import TestCase
 
-from project import app, db
+from project import create_app, db
 
+app = create_app()
 
 class BaseTestCase(TestCase):
     def create_app(self):
