+++ services/web/project/tests/test_config.py
@@ -4,8 +4,9 @@ import unittest
 from flask import current_app
 from flask_testing import TestCase
 
-from project import app
+from project import create_app
 
+app = create_app()
 
 class TestDevelopmentConfig(TestCase):
     def create_app(self):
